Api connection fixed

+++ App1/GreetMe_MVC/Models/Home/IndexModel.cs
@@ -11,7 +11,7 @@ namespace GreetMe_MVC.Models.Home
 
         public IndexModel(ViewDto viewDto)
         {
-            Id = viewDto.Id;
+            Id = (int)viewDto.Id;
             ViewName = viewDto.ViewName;
             HasCurrentDatetime = viewDto.HasCurrentDatetime;
             HasBirthday = viewDto.HasBirthday;
